phy: halbb: dig: skip out when registers don't exist
UBSAN reports errors due to shifts of 32 bits on u32s

The cause of this is mask == 0 passed to halbb_cal_bit_shift().

This comes from accessing registers that are not defined in
halbb_sdagc_follow_pagc_config() and halbb_dig_gain_para_init(), both
address and masks are zero.

Signed-off-by: Greg Whiteley <greg.whiteley@gmail.com>